Move `/S3.py` to `/storages/utils/S3.py`
This makes it possible to directly import the `S3.py` utility module
when django-storages is installed as pip-package, e.g.
`from storages.utils.S3 import CallingFormat`.
